requirements.txt
streamlit==1.38.0
langgraph==0.1.3
google-generativeai==0.8.2
beautifulsoup4==4.12.3
requests==2.32.3
serpapi==0.1.5
python-dotenv==1.0.1
langchain==0.2.12
wikipedia